requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Data Processing and Analysis
pandas>=1.5.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file support

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Image Processing and OCR
Pillow>=9.5.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pdf2image>=1.16.0

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0

# LLM Integration
together>=0.2.7
python-dotenv>=1.0.0

# Utility Libraries
requests>=2.31.0
python-multipart>=0.0.6  # For file uploads

# Optional: For enhanced functionality
scipy>=1.10.0  # Scientific computing
scikit-learn>=1.3.0  # Machine learning (optional)
wordcloud>=1.9.2  # Text visualization (optional)